+++ tests/test_util.py
@@ -4,8 +4,9 @@ import os
 import pathlib
 import re
 import unittest.mock as mock
-from datetime import datetime, date
+import datetime as dt
 from typing import List, Union
+
 import shapely.geometry
 import pytest
 
@@ -27,8 +28,8 @@ class TestRfc3339:
         assert "2020-03-17" == rfc3339.date("2020/03/17/12/34")
         assert "2020-03-17" == rfc3339.date("2020_03_17_12_34_56")
         assert "2020-03-17" == rfc3339.date("2020-03-17T12:34:56Z")
-        assert "2020-03-17" == rfc3339.date(date(2020, 3, 17))
-        assert "2020-03-17" == rfc3339.date(datetime(2020, 3, 17, 12, 34, 56))
+        assert "2020-03-17" == rfc3339.date(dt.date(2020, 3, 17))
+        assert "2020-03-17" == rfc3339.date(dt.datetime(2020, 3, 17, 12, 34, 56))
         assert "2020-03-17" == rfc3339.date((2020, 3, 17))
         assert "2020-03-17" == rfc3339.date([2020, 3, 17])
         assert "2020-03-17" == rfc3339.date(2020, 3, 17)
@@ -52,9 +53,9 @@ class TestRfc3339:
         assert "2020-03-17T12:34:00Z" == rfc3339.datetime("2020/03/17/12/34")
         assert "2020-03-17T12:34:56Z" == rfc3339.datetime("2020_03_17_12_34_56")
         assert "2020-03-17T12:34:56Z" == rfc3339.datetime("2020-03-17T12:34:56Z")
-        assert "2020-03-17T00:00:00Z" == rfc3339.datetime(date(2020, 3, 17))
+        assert "2020-03-17T00:00:00Z" == rfc3339.datetime(dt.date(2020, 3, 17))
         assert "2020-03-17T12:34:56Z" == rfc3339.datetime(
-            datetime(2020, 3, 17, 12, 34, 56)
+            dt.datetime(2020, 3, 17, 12, 34, 56)
         )
         assert "2020-03-17T00:00:00Z" == rfc3339.datetime((2020, 3, 17))
         assert "2020-03-17T00:00:00Z" == rfc3339.datetime([2020, 3, 17])
@@ -72,6 +73,18 @@ class TestRfc3339:
         assert "2020-09-17T12:34:56Z" == rfc3339.datetime(
             2020, "09", "17", "12", "34", 56
         )
+        assert "2020-03-17T12:34:56Z" == rfc3339.datetime(
+            dt.datetime(2020, 3, 17, 12, 34, 56, tzinfo=None)
+        )
+        assert "2020-03-17T12:34:56Z" == rfc3339.datetime(
+            dt.datetime(2020, 3, 17, 12, 34, 56, tzinfo=dt.timezone.utc)
+        )
+        assert "2020-03-17T12:34:56Z" == rfc3339.datetime(
+            dt.datetime(
+                *(2020, 3, 17, 12, 34, 56),
+                tzinfo=dt.timezone(offset=dt.timedelta(hours=0)),
+            )
+        )
 
     def test_normalize(self):
         assert "2020-03-17" == rfc3339.normalize("2020-03-17")
@@ -88,9 +101,21 @@ class TestRfc3339:
         assert "2020-03-17T12:34:56Z" == rfc3339.normalize(
             "2020-03-17T12:34:56.44546546Z"
         )
-        assert "2020-03-17" == rfc3339.normalize(date(2020, 3, 17))
+        assert "2020-03-17" == rfc3339.normalize(dt.date(2020, 3, 17))
         assert "2020-03-17T12:34:56Z" == rfc3339.normalize(
-            datetime(2020, 3, 17, 12, 34, 56)
+            dt.datetime(2020, 3, 17, 12, 34, 56)
+        )
+        assert "2020-03-17T12:34:56Z" == rfc3339.normalize(
+            dt.datetime(2020, 3, 17, 12, 34, 56, tzinfo=None)
+        )
+        assert "2020-03-17T12:34:56Z" == rfc3339.normalize(
+            dt.datetime(2020, 3, 17, 12, 34, 56, tzinfo=dt.timezone.utc)
+        )
+        assert "2020-03-17T12:34:56Z" == rfc3339.normalize(
+            dt.datetime(
+                *(2020, 3, 17, 12, 34, 56),
+                tzinfo=dt.timezone(offset=dt.timedelta(hours=0)),
+            )
         )
         assert "2020-03-17" == rfc3339.normalize((2020, 3, 17))
         assert "2020-03-17" == rfc3339.normalize([2020, 3, 17])
@@ -127,9 +152,9 @@ class TestRfc3339:
         assert formatter.date(None) is None
 
     def test_parse_date(self):
-        assert rfc3339.parse_date("2011-12-13") == date(2011, 12, 13)
+        assert rfc3339.parse_date("2011-12-13") == dt.date(2011, 12, 13)
         # `datetime.strptime` does not require leading zeros for month, day, hour, minutes, seconds
-        assert rfc3339.parse_date("0001-2-3") == date(1, 2, 3)
+        assert rfc3339.parse_date("0001-2-3") == dt.date(1, 2, 3)
 
     def test_parse_date_none(self):
         with pytest.raises(ValueError):
@@ -157,15 +182,21 @@ class TestRfc3339:
             rfc3339.parse_date(date)
 
     def test_parse_datetime(self):
-        assert rfc3339.parse_datetime("2011-12-13T14:15:16Z") == datetime(
+        assert rfc3339.parse_datetime("2011-12-13T14:15:16Z") == dt.datetime(
             2011, 12, 13, 14, 15, 16
         )
         # `datetime.strptime` is apparently case-insensitive about non-placeholder bits
-        assert rfc3339.parse_datetime("2011-12-13t14:15:16z") == datetime(
+        assert rfc3339.parse_datetime("2011-12-13t14:15:16z") == dt.datetime(
             2011, 12, 13, 14, 15, 16
         )
         # `datetime.strptime` does not require leading zeros for month, day, hour, minutes, seconds
-        assert rfc3339.parse_datetime("0001-2-3T4:5:6Z") == datetime(1, 2, 3, 4, 5, 6)
+        assert rfc3339.parse_datetime("0001-2-3T4:5:6Z") == dt.datetime(
+            1, 2, 3, 4, 5, 6
+        )
+        # Timezone handling
+        assert rfc3339.parse_datetime(
+            "2011-12-13T14:15:16Z", with_timezone=True
+        ) == dt.datetime(2011, 12, 13, 14, 15, 16, tzinfo=dt.timezone.utc)
 
     def test_parse_datetime_none(self):
         with pytest.raises(ValueError):
@@ -196,17 +227,20 @@ class TestRfc3339:
             rfc3339.parse_datetime(date)
 
     def test_parse_date_or_datetime(self):
-        assert rfc3339.parse_date_or_datetime("2011-12-13") == date(2011, 12, 13)
-        assert rfc3339.parse_date_or_datetime("0001-2-3") == date(1, 2, 3)
-        assert rfc3339.parse_date_or_datetime("2011-12-13T14:15:16Z") == datetime(
+        assert rfc3339.parse_date_or_datetime("2011-12-13") == dt.date(2011, 12, 13)
+        assert rfc3339.parse_date_or_datetime("0001-2-3") == dt.date(1, 2, 3)
+        assert rfc3339.parse_date_or_datetime("2011-12-13T14:15:16Z") == dt.datetime(
             2011, 12, 13, 14, 15, 16
         )
-        assert rfc3339.parse_date_or_datetime("2011-12-13t14:15:16z") == datetime(
+        assert rfc3339.parse_date_or_datetime("2011-12-13t14:15:16z") == dt.datetime(
             2011, 12, 13, 14, 15, 16
         )
-        assert rfc3339.parse_date_or_datetime("0001-2-3T4:5:6Z") == datetime(
+        assert rfc3339.parse_date_or_datetime("0001-2-3T4:5:6Z") == dt.datetime(
             1, 2, 3, 4, 5, 6
         )
+        assert rfc3339.parse_date_or_datetime(
+            "2011-12-13T14:15:16Z", with_timezone=True
+        ) == dt.datetime(2011, 12, 13, 14, 15, 16, tzinfo=dt.timezone.utc)
 
     def test_parse_date_or_datetime_none(self):
         with pytest.raises(ValueError):
@@ -353,19 +387,23 @@ class _Logger:
         self.logs.append(msg)
 
 
-def _fake_clock(times: List[Union[int, datetime]] = None):
+def _fake_clock(times: List[Union[int, dt.datetime]] = None):
     # Trick to have a "time" function that returns different times in subsequent calls
-    times = times or [datetime(2020, 3, 4, 5 + x, 2 * x, 1 + 3 * x, 1000) for x in range(0, 12)]
+    times = times or [
+        dt.datetime(2020, 3, 4, 5 + x, 2 * x, 1 + 3 * x, 1000) for x in range(0, 12)
+    ]
     return iter(times).__next__
 
 
 def test_timing_logger_custom():
     logger = _Logger()
     timing_logger = TimingLogger("Testing", logger=logger)
-    timing_logger._now = _fake_clock([
-        datetime(2019, 12, 12, 10, 10, 10, 10000),
-        datetime(2019, 12, 12, 11, 12, 13, 14141)
-    ])
+    timing_logger._now = _fake_clock(
+        [
+            dt.datetime(2019, 12, 12, 10, 10, 10, 10000),
+            dt.datetime(2019, 12, 12, 11, 12, 13, 14141),
+        ]
+    )
 
     with timing_logger:
         logger("Hello world")
@@ -380,10 +418,12 @@ def test_timing_logger_custom():
 def test_timing_logger_context_return():
     logger = _Logger()
     timing_logger = TimingLogger("Testing", logger=logger)
-    timing_logger._now = _fake_clock([
-        datetime(2019, 12, 12, 10, 10, 10, 0),
-        datetime(2019, 12, 12, 11, 12, 13, 0)
-    ])
+    timing_logger._now = _fake_clock(
+        [
+            dt.datetime(2019, 12, 12, 10, 10, 10, 0),
+            dt.datetime(2019, 12, 12, 11, 12, 13, 0),
+        ]
+    )
 
     with timing_logger as timer:
         logger("Hello world")
@@ -395,10 +435,12 @@ def test_timing_logger_context_return():
 def test_timing_logger_fail():
     logger = _Logger()
     timing_logger = TimingLogger("Testing", logger=logger)
-    timing_logger._now = _fake_clock([
-        datetime(2019, 12, 12, 10, 10, 10, 10000),
-        datetime(2019, 12, 12, 11, 12, 13, 14141)
-    ])
+    timing_logger._now = _fake_clock(
+        [
+            dt.datetime(2019, 12, 12, 10, 10, 10, 10000),
+            dt.datetime(2019, 12, 12, 11, 12, 13, 14141),
+        ]
+    )
 
     try:
         with timing_logger: